show charge and total charge in the cms in correct view,

+++ src/main/java/com/cob/billing/model/clinical/patient/CPTCode.java
@@ -9,5 +9,5 @@ public class CPTCode {
     private String serviceCode;
     private String modifier;
     private Integer unit;
-    private Float charge;
+    private double charge;
 }
